\name{nppack-package}
\alias{nppack-package}
\alias{nppack}
\docType{package}
\title{
\packageTitle{nppack}
}
\description{
\packageDescription{nppack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nppack}
\packageIndices{nppack}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{nppack}

Maintainer: \packageMaintainer{nppack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
